+++ src/main/java/com/cookiebytes/calmquest/counselor/CounselorRepository.java
@@ -3,6 +3,6 @@ package com.cookiebytes.calmquest.counselor;
 import com.cookiebytes.calmquest.student.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CounselorRepository extends JpaRepository<Counselor,Long> {
+public interface CounselorRepository extends JpaRepository<Counselor,Integer> {
 
 }
